+++ plugin/src/main/kotlin/org/gradle/github/dependencygraph/internal/util/PluginParameters.kt
@@ -1,29 +1,41 @@
 package org.gradle.github.dependencygraph.internal.util
 
-private const val ENV_VIA_SYS_PROP_PREFIX = "org.gradle.github.env."
+// TODO work out the best place for different constants
+const val PARAM_JOB_ID = "GITHUB_DEPENDENCY_GRAPH_JOB_ID"
+const val PARAM_JOB_CORRELATOR = "GITHUB_DEPENDENCY_GRAPH_JOB_CORRELATOR"
+const val PARAM_GITHUB_REF = "GITHUB_REF"
+const val PARAM_GITHUB_SHA = "GITHUB_SHA"
 
-const val ENV_DEPENDENCY_GRAPH_JOB_ID = "GITHUB_DEPENDENCY_GRAPH_JOB_ID"
-const val ENV_DEPENDENCY_GRAPH_JOB_CORRELATOR = "GITHUB_DEPENDENCY_GRAPH_JOB_CORRELATOR"
-const val ENV_GITHUB_REF = "GITHUB_REF"
-const val ENV_GITHUB_SHA = "GITHUB_SHA"
+const val PARAM_INCLUDE_PROJECTS = "DEPENDENCY_GRAPH_INCLUDE_PROJECTS"
+const val PARAM_INCLUDE_CONFIGURATIONS = "DEPENDENCY_GRAPH_INCLUDE_CONFIGURATIONS"
 
-const val ENV_DEPENDENCY_GRAPH_REPORT_DIR = "GITHUB_DEPENDENCY_GRAPH_REPORT_DIR"
+const val PARAM_REPORT_DIR = "GITHUB_DEPENDENCY_GRAPH_REPORT_DIR"
 
 /**
  * Environment variable should be set to the workspace directory that the Git repository is checked out in.
  * This is used to determine relative path to build files referenced in the dependency graph.
  */
-const val ENV_GITHUB_WORKSPACE = "GITHUB_WORKSPACE"
+const val PARAM_GITHUB_WORKSPACE = "GITHUB_WORKSPACE"
 
 class PluginParameters {
     fun load(envName: String, default: String? = null): String {
-        return System.getProperty(ENV_VIA_SYS_PROP_PREFIX + envName)
+        return System.getProperty(sysPropName(envName))
             ?: System.getenv()[envName]
             ?: default
             ?: throwEnvironmentVariableMissingException(envName)
     }
 
+    fun loadOptional(envName: String): String? {
+        return System.getProperty(sysPropName(envName))
+            ?: System.getenv()[envName]
+    }
+
+    private fun sysPropName(envName: String): String {
+        return envName.toLowerCase().replace('_', '-')
+    }
+
     private fun throwEnvironmentVariableMissingException(variable: String): Nothing {
-        throw IllegalStateException("The configuration parameter '$variable' must be set, but it was not.")
+        throw IllegalStateException("The configuration parameter '$variable' must be set: " +
+            "set an environment variable, or '-D${sysPropName(variable)}=value' on the command-line.")
     }
 }