+++ release/changes.md
@@ -1,3 +1,5 @@
 Preliminary release of the github dependency graph plugin.
 
 - Reduce verbosity of log messages
+- Use a custom task for resolving all project dependencies, instead of calling the "dependencies" task
+- Which Gradle Configurations contribute to the dependency graph can be filtered by project or configuration name